Added a regression test for issue 11678 (spack compiler find broken)
